Create Candy Machine Program
A new candy machine is bought for the gym, but it is not working properly. The Machine sells candies, chips, gum, and cookies. In this programming problem, you will create a program that:

Show the customer the different products sold by the candy machine.
Let the customer make the selection.
Show the customer the cost of the item selected.
Accept money from the customer.
Release the item.
Input / Output:
Input: The item selection and the cost of the item
Output: The selected item.